fix: remove complexity ratchet baseline exemption
