+++ lib/complexity_ratchet.rb
@@ -1,10 +1,9 @@
 # frozen_string_literal: true
 
 require "date"
-require "json"
-require "open3"
 require "prism"
 require "yaml"
+require_relative "complexity_ratchet/baseline_operations"
 
 # Entity-level monotonic complexity ratchet.
 #
@@ -436,88 +435,6 @@ module ComplexityRatchet
     end
   end
 
-  # Runs RuboCop's Metrics department and folds the offenses into
-  # {entity key => measured value}.
-  class Measurement
-    def self.call(root:, config: CONFIG_PATH)
-      new(root: root, config: config).call
-    end
-
-    def initialize(root:, config: CONFIG_PATH)
-      @root = root
-      @config = config
-    end
-
-    def call
-      fold(run_rubocop)
-    end
-
-    # Split out from #call so the folding — which is where the entity keys and
-    # the max-vs-count rules live — can be tested against a fixture payload
-    # instead of a two-second RuboCop run.
-    def fold(payload)
-      payload.fetch("files").each_with_object({}) do |file, acc|
-        offenses = file["offenses"]
-        next if offenses.nil? || offenses.empty?
-
-        absorb(acc, relative_path(file["path"]), offenses)
-      end
-    end
-
-    private
-
-    attr_reader :root, :config
-
-    def run_rubocop
-      command = [ "bundle", "exec", "rubocop", "-c", config, "--ignore-disable-comments",
-                  "--only", "Metrics", "--format", "json", "--no-color" ]
-      stdout, stderr, status = Open3.capture3(*command, chdir: root)
-      # RuboCop exits 1 whenever offenses exist, which is the normal case here.
-      # Only a missing JSON document means the run itself failed.
-      raise Error, "rubocop failed (#{status.exitstatus}): #{stderr}" if stdout.strip.empty?
-
-      JSON.parse(stdout)
-    end
-
-    def absorb(acc, path, offenses)
-      source = File.read(File.join(root, path))
-      entities = EntityMap.for(source)
-      lines = source.lines
-      offenses.sort_by { |offense| [ offense.dig("location", "start_line") || offense.dig("location", "line"), offense.dig("location", "last_line") || 0 ] }.each do |offense|
-        line = offense.dig("location", "start_line") || offense.dig("location", "line")
-        last_line = offense.dig("location", "last_line")
-        key = [ path, offense["cop_name"], entities[line, last_line] || fallback_name(entities, lines, line, last_line) ].join(SEPARATOR)
-        record(acc, key, offense["message"])
-      end
-    end
-
-    # Metrics/BlockNesting and friends point at a bare statement rather than a
-    # definition. The normalised source line and enclosing scope identify the
-    # usual case; include below-budget twins in the ordinal population too.
-    def fallback_name(entities, lines, line, last_line)
-      text = lines[line - 1].to_s.strip.gsub(/\s+/, " ")
-      text = "#{text[0, 97]}..." if text.length > 100
-      base = [ entities.enclosing_path(line, last_line), "~#{text}" ].compact.join
-      ordinal = entities.fallback_ordinal(line, last_line, text)
-      ordinal && ordinal > 1 ? "#{base}[fallback:#{ordinal}]" : base
-    end
-
-    def record(acc, key, message)
-      measured = message[VALUE_PATTERN, 1]
-      if measured
-        acc[key] = ComplexityRatchet.normalize([ acc[key].to_f, measured.to_f ].max)
-      else
-        acc[key] = acc[key].to_i + 1
-      end
-    end
-
-    def relative_path(path)
-      absolute = File.expand_path(path)
-      prefix = "#{File.expand_path(root)}/"
-      absolute.start_with?(prefix) ? absolute.delete_prefix(prefix) : path
-    end
-  end
-
   # A single reason the ratchet is unhappy. `blocking` separates hard failures
   # from advisory notes so an unused waiver cannot fail a build on its own.
   Problem = Struct.new(:kind, :key, :message, :blocking, keyword_init: true) do
@@ -624,39 +541,9 @@ module ComplexityRatchet
     end
   end
 
+  extend BaselineOperations
+
   class << self
-    # RuboCop prints AbcSize as a float; keep whole numbers as integers so the
-    # baseline diff reads 38 rather than 38.0.
-    def normalize(value)
-      rounded = value.round(2)
-      rounded == rounded.to_i ? rounded.to_i : rounded
-    end
-
-    def load_baseline(path) = File.exist?(path) ? flatten(YAML.safe_load_file(path) || {}).tap { |entries| validate_baseline_values!(entries) } : {}
-
-    def dump_baseline(path, entries)
-      File.write(path, BASELINE_HEADER + YAML.dump(nest(entries), line_width: -1))
-    end
-
-    # Baselines are nested on disk, but flattened to `path | cop | entity` in memory.
-    # Nesting prevents YAML from expanding long keys into its multi-line explicit form.
-    def nest(entries)
-      entries.sort.each_with_object({}) do |(key, value), acc|
-        file, cop, entity = key.split(SEPARATOR, 3)
-        ((acc[file] ||= {})[cop] ||= {})[entity] = value
-      end
-    end
-
-    def flatten(nested)
-      nested.each_with_object({}) do |(file, cops), acc|
-        cops.each do |cop, entities|
-          entities.each { |entity, value| acc[[ file, cop, entity ].join(SEPARATOR)] = value }
-        end
-      end
-    end
-
-    def validate_baseline_values!(entries) = entries.each { |key, value| raise Error, "baseline entry #{key.inspect} must be a finite nonnegative number (got #{value.inspect})" unless value.is_a?(Numeric) && value.finite? && value >= 0 }
-
     def load_waivers(path)
       return [] unless File.exist?(path)
 
@@ -1050,3 +937,5 @@ module ComplexityRatchet
     end
   end
 end
+
+require_relative "complexity_ratchet/measurement"
